[svn r398] Fixed the regressed behavior in the Contour Finder.  The issue came about in that I was directly accessing the ->data[] attributes of the PyArrayObjects; I am not sure why this was not working, but it wasn't.  I switched to PyArray_GETPTR1.  Now I get correct behavior for my regression tests, and all the units tests passed.  Additionally, I cleaned up some 'type' stuff in the C code and in the ContourFinder.py file.
Added more unit tests for datacubes and contour finding.  Also made it so that
you could flush back a data field from a datacube even if that field is new.

--HG--
branch : trunk

+++ yt/lagos/ContourFinder.py
@@ -102,7 +102,7 @@ def identify_contours(data_source, field, min_val, max_val):
             kk = na.arange(cur_max_id, cur_max_id-local_ind[0].size, -1)
             cg["tempContours"][local_ind] = kk[:]
             cur_max_id -= local_ind[0].size
-        fd = cg["tempContours"]
+        fd = cg["tempContours"].astype('int64')
         fd_original = fd.copy()
         xi_u,yi_u,zi_u = na.where(fd > -1)
         cor_order = na.argsort(-1*fd[(xi_u,yi_u,zi_u)])
@@ -110,7 +110,7 @@ def identify_contours(data_source, field, min_val, max_val):
         yi = yi_u[cor_order]
         zi = zi_u[cor_order]
         PointCombine.FindContours(fd, xi, yi, zi)
-        cg["tempContours"] = fd.copy()
+        cg["tempContours"] = fd.copy().astype('float64')
         cg.flush_data("tempContours")
         my_queue.add(cg._grids)
         force_ind = na.unique(cg["GridIndices"][na.where(
